+++ CHANGELOG.md
@@ -5,14 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-## Fixed
+### Fixed
 - The readonly-keyword rule now properly checks for readonly modifier of `class` property declarations. See [49](https://github.com/jonaskello/tslint-immutable/issues/49) and PR [#50](https://github.com/jonaskello/tslint-immutable/pull/50).
 
 ### Added
 - New options `ignore-local`, and `ignore-prefix` for the `no-let` rule. See [#32](https://github.com/jonaskello/tslint-immutable/issues/39), also requested in [#43](https://github.com/jonaskello/tslint-immutable/issues/43). See PR [#48](https://github.com/jonaskello/tslint-immutable/pull/48).
 
 ## [v4.2.1] - 2017-09-21
-## Fixed
+### Fixed
 - The readonly-array rule with ignore-local option does not work within `class`. See [45](https://github.com/jonaskello/tslint-immutable/issues/45).
 
 ## [v4.2.0] - 2017-09-14